fixes for unicode input
when examparser.py is running in python 2.7 (for the editor), we need the literal '\ufeff' to be unicode. So import unicode_strings from __future__.
Detach the STDIN for numbas.py so it reads bytes, which can then be decoded as utf-8. Without doing this, the stuff from stdin was always forced to ASCII.